refactor: organize components and projects structure, renamed /customWidgets into /components
