--- lib/algorithms/navigationObservers.dart
@@ -1,4 +1,0 @@
-import 'package:flutter/material.dart';
-
-final RouteObserver<ModalRoute<void>> routeObserver =
-    RouteObserver<ModalRoute<void>>();
