+++ lib/components/floatingSearch.dart
