+++ lib/components/noteShortcut.dart
