+++ lib/components/search_note_dialog.dart
